pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ad_platform_env"
version = "0.1.0"
description = "Multi-task ad platform RL environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gymnasium>=1.2.3",
    "openenv-core>=0.2.2",
    "numpy>=1.24.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "ad_platform_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["ad_platform_env", "ad_platform_env.server"]
package-dir = { "ad_platform_env" = ".", "ad_platform_env.server" = "server" }

